Adds function to change text in "display" div
This commit adds the function updateDisplay(newValue) which takes a
number parameter, converts it to a string with maximum 1 decimal
place, and changes the innerText of the "display" class div to that
string.

The commit also adds a global const variable for the "display" class
div, and adds comments to previously committed js code.