Adds functionality to calculator number buttons
This commit adds the function numberClick(num) which takes a string
parameter which is a single character corresponding to either a
digit or period. If the string is a digit, and the display is able to
take further input (i.e., is less than 8 digits and does not have any
decimal places), it appends the digit to the end of the current
display value and calls updateDisplay with the new value.

This commit also adds an onclick to each of the buttons of class
"number", which calls numberClick with the innerText of that button.
The result is that clicking the number buttons on the webpage updates
the display until it reaches 8 digits.

This commit also adds a function updateDisplayDecimal(), which
does not currently do anything. If the parameter for numberClick is
a period, it calls updateDisplayDecimal and returns.