Adds functionality to the operator buttons
This commit adds a function operatorClick(operator) which takes a
string parameter which is a calculator operator (+, -, *, /, =) and
evaluates and updates the calculator display. This function is then
set as the onclick action for each button of class operator, with the
innerText as the parameter used.

In order to do this, the js file also makes some updates to the
global variables. Instead of two operands, it has just one variable,
"total", which represents the total amount stored in the "calculator
memory". It also uses the global variable currentOperator to store
the most recent non-equals operator clicked in "calculator memory".
Finally, there is a global variable inputRestart which is boolean and
is true when the next number button press should reset the current
value in the calculator display.

This commit also slightly re-organizes the code in the JS file into
different sections for different purposes (e.g., for display updates,
for enabling number buttons, for enabling operator buttons).

NB: I am surprised that I have not found any bugs in testing this
functionality, and I kind of think there must be something wrong that
I haven't caught yet.