Adds functionality to the +/- button
This commit adds a function plusMinus() which multiplies the value
in the display by -1, so long as it is a number (i.e., not "ERROR")
and it is a user-entered number (i.e., not the displayed result of a
calculation). The plusMinus function is added as the onclick proeprty
of the button with ID "+-".