Adds functionality to the Del button
This commit adds the function delButton() which deletes the last
user-entered number or decimal point from the displayed value. It
does not function if the calculator is in an error state or the value
displayed is the result of a calculation, rather than user-entry.
This function is set as the onclick value for the button with ID
"del".