Adds keyboard numpad control to the calculator
This commit enables control of the calculator through the keyboard
numpad, as well as the Del key for the delete button. This is
accomplished through event listeners for keydown and keyup.

This commit adds two functions, handleKeypress(event) and
handleKeyRelease(event). Both take an event as a parameter, and check
the keyCode of the event to see if it corresponds to the data-keycode
property of one of the buttons. If so, the former changes the
border style of the button to inset and calls its onclick function;
the latter resets the border style of the button to outset.

This commit also adds a class "keyboardPress" to styles.css, which is
the class added and removed in the handleKeypress and
handleKeyRelease functions to change the border style of the button.

This commit also updates the data-keycode property of the "=" button
so that it is triggered by numpad enter.