Fixes a bug when using both numpad and mouse
Previously, when using both numpad to control the calculator and
clicking directly on buttons numpad enter would re-click the last
button clicked, which screwed up the calculator function. This commit
adds event.preventDefault() calls to the keypress event handler
functions to avoid this.

There is still an issue where pressing numpad keys changes the border
formatting of the last button clicked - have not figured out how to
fix that yet.